+++ common/display.h
@@ -5,7 +5,7 @@
 
 void sevenSeg_scrollText(const char * text, int8_t times); // -1 -> infinite / 0 -> off
 void sevenSeg_setAscii(char left, char right);
-void sevenSeg_setNumber(int8_t n);
+void sevenSeg_setNumber(int32_t n);
 
 int led_getOn(p600LED_t led);
 int led_getBlinking(p600LED_t led);
